Added basic JS to frontend.
Added gems to load JS dependencies for todo jQuery app.
Added html response for showing a list, which just serves the jQuery 
app.
Updated the layout to include scripts at the bottom of the body tag and
with an improved title.
Added a root route that redirects to list 1.
Added source for Todo jQuery app and CSS.