Finish basic Todo functionality.
Changed Todos to use the client-generated UUIDs as the primary key.
Added a method to the TodoList to fully replace the set of todos based
on the attributes passed in.
Added back jquery_ujs for the free CSRF protection added to AJAX
requests.
Added push and fetch methods to the TodoMVC app for simple saving on
render and getting initial set of Todos.  Also, wait to render initially
until todos are fetched.